Convert inspection into intention
