+++ resources/META-INF/plugin.xml
@@ -675,10 +675,6 @@
         <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.LatexFigureNotReferencedInspection"
                          groupName="LaTeX" displayName="Figure not referenced"
                          enabledByDefault="true"/>
-        <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.LatexVerbatimInspection"
-                         groupName="LaTeX" displayName="Verbatim environments might break formatter and/or parser"
-                         shortName="LatexUseOfVerbatimEnvironment"
-                         enabledByDefault="true"/>
         <localInspection language="Latex" implementationClass="nl.hannahsten.texifyidea.inspections.latex.LatexMultipleGraphicsPathInspection"
                          groupName="LaTeX" displayName="Multiple definition of graphicspath"
                          enabledByDefault="true"/>
@@ -713,6 +709,36 @@
         </intentionAction>
 
         <intentionAction>
+            <className>nl.hannahsten.texifyidea.intentions.LatexInlineDisplayToggle</className>
+            <category>LaTeX</category>
+            <descriptionDirectoryName>LatexInlineDisplayToggle</descriptionDirectoryName>
+        </intentionAction>
+
+        <intentionAction>
+            <className>nl.hannahsten.texifyidea.intentions.LatexInsertFormatterCommentsIntention</className>
+            <category>LaTeX</category>
+            <descriptionDirectoryName>LatexInsertFormatterComments</descriptionDirectoryName>
+        </intentionAction>
+
+        <intentionAction>
+            <className>nl.hannahsten.texifyidea.intentions.LatexLabelDefiningNewCommand</className>
+            <category>LaTeX</category>
+            <descriptionDirectoryName>LatexLabelDefiningNewCommand</descriptionDirectoryName>
+        </intentionAction>
+
+        <intentionAction>
+            <className>nl.hannahsten.texifyidea.intentions.LatexLeftRightParenthesesIntention</className>
+            <category>LaTeX</category>
+            <descriptionDirectoryName>LatexLeftRightParentheses</descriptionDirectoryName>
+        </intentionAction>
+
+        <intentionAction>
+            <className>nl.hannahsten.texifyidea.intentions.LatexMathToggle</className>
+            <category>LaTeX</category>
+            <descriptionDirectoryName>LatexMathToggle</descriptionDirectoryName>
+        </intentionAction>
+
+        <intentionAction>
             <className>nl.hannahsten.texifyidea.intentions.LatexMoveSectionToFileIntention</className>
             <category>LaTeX</category>
             <descriptionDirectoryName>LatexMoveSectionToFile</descriptionDirectoryName>
@@ -725,39 +751,15 @@
         </intentionAction>
 
         <intentionAction>
-            <className>nl.hannahsten.texifyidea.intentions.LatexUnpackUsepackageIntention</className>
-            <category>LaTeX</category>
-            <descriptionDirectoryName>LatexUnpackUsepackage</descriptionDirectoryName>
-        </intentionAction>
-
-        <intentionAction>
-            <className>nl.hannahsten.texifyidea.intentions.LatexLeftRightParenthesesIntention</className>
-            <category>LaTeX</category>
-            <descriptionDirectoryName>LatexLeftRightParentheses</descriptionDirectoryName>
-        </intentionAction>
-
-        <intentionAction>
-            <className>nl.hannahsten.texifyidea.intentions.LatexInlineDisplayToggle</className>
-            <category>LaTeX</category>
-            <descriptionDirectoryName>LatexInlineDisplayToggle</descriptionDirectoryName>
-        </intentionAction>
-
-        <intentionAction>
-            <className>nl.hannahsten.texifyidea.intentions.LatexMathToggle</className>
-            <category>LaTeX</category>
-            <descriptionDirectoryName>LatexMathToggle</descriptionDirectoryName>
-        </intentionAction>
-
-        <intentionAction>
             <className>nl.hannahsten.texifyidea.intentions.LatexToggleSmartQuotesIntention</className>
             <category>LaTeX</category>
             <descriptionDirectoryName>LatexToggleSmartQuotes</descriptionDirectoryName>
         </intentionAction>
 
         <intentionAction>
-            <className>nl.hannahsten.texifyidea.intentions.LatexLabelDefiningNewCommand</className>
+            <className>nl.hannahsten.texifyidea.intentions.LatexUnpackUsepackageIntention</className>
             <category>LaTeX</category>
-            <descriptionDirectoryName>LatexLabelDefiningNewCommand</descriptionDirectoryName>
+            <descriptionDirectoryName>LatexUnpackUsepackage</descriptionDirectoryName>
         </intentionAction>
     </extensions>
 </idea-plugin>
