--- resources/inspectionDescriptions/LatexUseOfVerbatimEnvironment.html
@@ -1,16 +1,0 @@
-<html>
-<body>
-A verbatim environment might break TeXiFy's formatter and/or parser.
-<!-- tooltip end -->
-<p>
-    There are two quick fixes available:
-    <li>Move the verbatim environment to a new file.
-    This new file will contain just the verbatim environment, and the formatter will be disabled on this file.
-    Choose this fix if the verbatim environment caused parse errors.</li>
-    <li>Surround the verbatim environment with comments that disable the formatter.</li>
-</p>
-<p>
-    Bottom line: be careful with these statements and use a different command name where possible, unless you are very certain of what you are doing.
-</p>
-</body>
-</html>
