+++ resources/intentionDescriptions/LatexInsertFormatterComments/after.tex.template
@@ -1,0 +1,9 @@
+% @formatter:off
+\begin{lstlisting}[language=Kotlin]
+fun Int?.ifPositiveAddTwo(): Int =
+        this?.let {
+            if (this >= 0) this + 2
+            else this
+        } ?: 0
+\end{lstlisting}
+% @formatter:on
