+++ resources/intentionDescriptions/LatexInsertFormatterComments/description.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+Insert comments to disable the formatter.
+</body>
+</html>
