+++ src/nl/hannahsten/texifyidea/grammar/LatexLexer.flex
@@ -214,8 +214,7 @@ ANY_CHAR=.|\n
 // because that will confuse the formatter because it will see the next line as being on this line
 \\\n                 { return com.intellij.psi.TokenType.WHITE_SPACE; }
 
-// todo syntax highlighting on \end
-// todo remove formatter inspections?
+// todo syntax highlighting on \end. Will have to be fixed in syntax highlighter as lexer cannot distinguish \ends
 // todo insert label quickfix broken
 
 "*"                  { return STAR; }
